refactor: remove committed d.ts files and relocate preload types to renderer
